Xaml separation and media sample
Had to separate XAML library into two separate projects because they
required different references for the Behaviors SDK. Though they are two
projects, they produce the same DLL name. Will need to keep this in mind
when producing NuGet packages. MediaPage sample is for the most part
done. It supports Phone and Tablet in both Landscape and Portrait.
Performance is not good on phone but I can't tell why.